# src/main/resources/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Mithra Direct App",
    "description": "API documentation for the subscription services",
    "contact": {
      "name": "Swamy Kunta",
      "url": "https://github.com/s713278/ECommerceApp/tree/master",
      "email": "swamy.kunta@gmail.com"
    },
    "license": {
      "name": "License",
      "url": "/"
    },
    "version": "v1"
  },
...
  "servers": [
    {
      "url": "http://localhost:8080/api",
      "description": "Local Env"
    },
    {
      "url": "http://172.105.33.153:8080/api",
      "description": "Dev Env"
    }
  ],
  "paths": {
...
    "/v1/vendors/{vendorId}": {
      "put": {
        "tags": [
          "2. Vendors/Product Listing"
        ],
        "operationId": "updateStore",
        "parameters": [
          {
            "name": "vendorId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VendorDetailsDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponseVendorDetailsDTO"
                }
              }
            }
          }
        },
        "security": [
          {
            "E-Commerce Application": []
          }
        ]
      },
      "delete": {
        "tags": [
          "2. Vendors/Product Listing"
        ],
        "operationId": "deleteStore",
        "parameters": [
          {
            "name": "vendorId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponseString"
                }
              }
            }
          }
        },
        "security": [
          {
            "E-Commerce Application": []
          }
        ]
      }
    },
    "/v1/vendors/{vendorId}/orders/{orderId}": {
      "get": {
        "tags": [
          "5. Vendor Management"
        ],
        "summary": "Get a specific order",
        "operationId": "getOrdersById",
        "parameters": [
          {
            "name": "orderId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponseOrderDTO"
                }
              }
            }
          }
        },
        "security": [
          {
            "E-Commerce Application": []
          }
        ]
      },
      "put": {
        "tags": [
          "5. Vendor Management"
        ],
        "summary": "Update and existing order",
        "operationId": "updateOrder",
        "parameters": [
          {
            "name": "orderId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OrderUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponseOrderUpdateResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "E-Commerce Application": []
          }
        ]
      }
    },
...
    "/v1/vendors/{vendorId}/orders/": {
      "get": {
        "tags": [
          "5. Vendor Management"
        ],
        "summary": "Get all orders",
        "operationId": "getOrdersByVendor",
        "parameters": [
          {
            "name": "vendorId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponseListOrderDTO"
                }
              }
            }
          }
        },
        "security": [
          {
            "E-Commerce Application": []
          }
        ]
      },
      "post": {
        "tags": [
          "5. Vendor Management"
        ],
        "operationId": "placeOrder",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OrderRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponseOrderDTO"
                }
              }
            }
          }
        },
        "security": [
          {
            "E-Commerce Application": []
          }
        ]
      }
    },
    "/v1/vendors/": {
      "get": {
        "tags": [
          "2. Vendors/Product Listing"
        ],
        "summary": "All vendors listing",
        "operationId": "fetchAllActiveVendors",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponseObject"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "2. Vendors/Product Listing"
        ],
        "operationId": "createStore",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VendorDetailsDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponseVendorDetailsDTO"
                }
              }
            }
          }
        },
        "security": [
          {
            "E-Commerce Application": []
          }
        ]
      }
    },
...
    "/v1/vendors/{zipCode}": {
      "get": {
        "tags": [
          "2. Vendors/Product Listing"
        ],
        "summary": "Vendors listing by zipcode",
        "operationId": "fetchAllActiveVendorsByZipCode",
        "parameters": [
          {
            "name": "zipCode",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponseObject"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/vendors/{vendorId}/orders/public/users/{userId}/orders": {
      "get": {
        "tags": [
          "5. Vendor Management"
        ],
        "operationId": "getOrdersByUser",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/OrderDTO"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "E-Commerce Application": []
          }
        ]
      }
    },
    "/v1/vendors/{vendorId}/orders/public/users/{emailId}/orders/{orderId}": {
      "get": {
        "tags": [
          "5. Vendor Management"
        ],
        "operationId": "getOrderByUser",
        "parameters": [
          {
            "name": "emailId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "orderId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/OrderDTO"
                }
              }
            }
          }
        },
        "security": [
          {
            "E-Commerce Application": []
          }
        ]
      }
    },
    "/v1/vendors/{vendorId}/orders/admin/orders": {
      "get": {
        "tags": [
          "5. Vendor Management"
        ],
        "operationId": "getAllOrders",
        "parameters": [
          {
            "name": "pageNumber",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 2
            }
          },
          {
            "name": "sortBy",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "totalAmount"
            }
          },
          {
            "name": "sortOrder",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "asc"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/OrderResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "E-Commerce Application": []
          }
        ]
      }
    },
...
    "/v1/users/{userId}/vendor/{vendorId}": {
      "get": {
        "tags": [
          "5. User Management"
        ],
        "summary": "Fetch Subscriptions By Vendor and User",
        "operationId": "fetchSubsByUserAndVendor",
        "parameters": [
          {
            "name": "vendorId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponseObject"
                }
              }
            }
          }
        },
        "security": [
          {
            "E-Commerce Application": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "VendorDetailsDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "status": {
            "type": "string",
            "enum": [
              "NEW",
              "ACTIVE",
              "DELETED",
              "SUSPENDED"
            ]
          },
          "contact_number": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "support_number": {
            "type": "string"
          },
          "verification_status": {
            "type": "string",
            "enum": [
              "PENDING",
              "APPROVED",
              "REJECTED"
            ]
          }
        }
      },
      "APIResponseVendorDetailsDTO": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "description": "Response message",
            "example": "Customer registered successfully"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          },
          "success": {
            "type": "boolean"
          },
          "status": {
            "type": "integer",
            "description": "HTTP Status Code",
            "format": "int32",
            "example": 201
          },
          "data": {
            "$ref": "#/components/schemas/VendorDetailsDTO"
          }
        }
      },
      "OrderUpdateRequest": {
        "type": "object",
        "properties": {
          "old_status": {
            "type": "string",
            "description": "Old Status",
            "example": "CREATED/PROCESSING/SHIPPED/DELIVERED",
            "enum": [
              "PENDING",
              "SCHEDULED",
              "IN_PROCESS",
              "DELIVERED",
              "CANCELED"
            ]
          },
          "new_status": {
            "type": "string",
            "description": "New Status",
            "example": "PROCESSING",
            "enum": [
              "PENDING",
              "SCHEDULED",
              "IN_PROCESS",
              "DELIVERED",
              "CANCELED"
            ]
          }
        }
      },
      "APIResponseOrderUpdateResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "description": "Response message",
            "example": "Customer registered successfully"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          },
          "success": {
            "type": "boolean"
          },
          "status": {
            "type": "integer",
            "description": "HTTP Status Code",
            "format": "int32",
            "example": 201
          },
          "data": {
            "$ref": "#/components/schemas/OrderUpdateResponse"
          }
        }
      },
      "OrderUpdateResponse": {
        "type": "object",
        "properties": {
          "order_id": {
            "type": "integer",
            "format": "int64"
          },
          "old_satus": {
            "type": "string",
            "enum": [
              "PENDING",
              "SCHEDULED",
              "IN_PROCESS",
              "DELIVERED",
              "CANCELED"
            ]
          },
          "new_staus": {
            "type": "string",
            "enum": [
              "PENDING",
              "SCHEDULED",
              "IN_PROCESS",
              "DELIVERED",
              "CANCELED"
            ]
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          }
        },
        "description": "Data"
      },
...
      "APIResponseObject": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "description": "Response message",
            "example": "Customer registered successfully"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          },
          "success": {
            "type": "boolean"
          },
          "status": {
            "type": "integer",
            "description": "HTTP Status Code",
            "format": "int32",
            "example": 201
          },
          "data": {
            "type": "object",
            "description": "Data"
          }
        }
      },
      "AddressDTO": {
        "type": "object",
        "properties": {
          "address1": {
            "type": "string",
            "description": "Address1",
            "example": "73 Owen Lane"
          },
          "address2": {
            "type": "string"
          },
          "city": {
            "type": "string",
            "description": "City",
            "example": "Fort Myers"
          },
          "state": {
            "type": "string",
            "description": "State",
            "example": "Florida"
          },
          "country": {
            "type": "string",
            "description": "Country",
            "example": "United States"
          },
          "pincode": {
            "type": "string",
            "description": "Zipcode",
            "example": "33901"
          }
        }
      },
      "OrderRequest": {
        "type": "object",
        "properties": {
          "user_id": {
            "type": "integer",
            "format": "int64"
          },
          "cart_id": {
            "type": "integer",
            "format": "int64"
          },
          "payment_details": {
            "$ref": "#/components/schemas/PaymentDTO"
          },
          "shipping_details": {
            "$ref": "#/components/schemas/ShippingDTO"
          }
        }
      },
      "PaymentDTO": {
        "type": "object",
        "properties": {
          "credit_card": {
            "type": "string",
            "description": "Creditcard Number",
            "example": "XXXX-XXXX-XXXX-1280"
          },
          "expiry_date": {
            "type": "string",
            "description": "Exp Date",
            "example": "12/2030"
          },
          "payment_method": {
            "type": "string",
            "description": "Payment Method",
            "example": "IN_STORE_PAYMENT"
          },
          "billing_address": {
            "$ref": "#/components/schemas/AddressDTO"
          }
        }
      },
      "ShippingDTO": {
        "type": "object",
        "properties": {
          "address": {
            "$ref": "#/components/schemas/AddressDTO"
          },
          "shipping_type": {
            "type": "string",
            "description": "Shipping Type",
            "example": "INSTORE_PICKUP"
          }
        }
      },
      "APIResponseOrderDTO": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "description": "Response message",
            "example": "Customer registered successfully"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          },
          "success": {
            "type": "boolean"
          },
          "status": {
            "type": "integer",
            "description": "HTTP Status Code",
            "format": "int32",
            "example": 201
          },
          "data": {
            "$ref": "#/components/schemas/OrderDTO"
          }
        }
      },
      "OrderDTO": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          },
          "order_id": {
            "type": "integer",
            "format": "int64"
          },
          "order_date": {
            "type": "string",
            "format": "date-time"
          },
          "store_name": {
            "type": "string"
          },
          "order_items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OrderItemDTO"
            }
          },
          "payment_details": {
            "$ref": "#/components/schemas/PaymentDTO"
          },
          "total_amount": {
            "type": "number",
            "format": "double"
          },
          "order_status": {
            "type": "string"
          },
          "shipping_details": {
            "$ref": "#/components/schemas/ShippingDTO"
          },
          "sub_total": {
            "type": "number",
            "format": "double"
          },
          "federal_tax": {
            "type": "number",
            "format": "double"
          },
          "state_tax": {
            "type": "number",
            "format": "double"
          }
        },
        "description": "Data"
      },
      "OrderItemDTO": {
        "type": "object",
        "properties": {
          "quantity": {
            "type": "integer",
            "format": "int32"
          },
          "discount": {
            "type": "number",
            "format": "double"
          },
          "amount": {
            "type": "number",
            "format": "double"
          },
          "unit_price": {
            "type": "number",
            "format": "double"
          },
          "state_tax": {
            "type": "number",
            "format": "double"
          },
          "federal_tax": {
            "type": "number",
            "format": "double"
          }
        }
      },
...
      "OrderResponse": {
        "type": "object",
        "properties": {
          "content": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OrderDTO"
            }
          },
          "pageNumber": {
            "type": "integer",
            "format": "int32"
          },
          "pageSize": {
            "type": "integer",
            "format": "int32"
          },
          "totalElements": {
            "type": "integer",
            "format": "int64"
          },
          "totalPages": {
            "type": "integer",
            "format": "int32"
          },
          "lastPage": {
            "type": "boolean"
          }
        }
      },
      "APIResponseListOrderDTO": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "description": "Response message",
            "example": "Customer registered successfully"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          },
          "success": {
            "type": "boolean"
          },
          "status": {
            "type": "integer",
            "description": "HTTP Status Code",
            "format": "int32",
            "example": 201
          },
          "data": {
            "type": "array",
            "description": "Data",
            "items": {
              "$ref": "#/components/schemas/OrderDTO"
            }
          }
        }
      },
...
      "APIResponseString": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "description": "Response message",
            "example": "Customer registered successfully"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          },
          "success": {
            "type": "boolean"
          },
          "status": {
            "type": "integer",
            "description": "HTTP Status Code",
            "format": "int32",
            "example": 201
          },
          "data": {
            "type": "string",
            "description": "Data"
          }
        }
      }
    },
    "securitySchemes": {
      "E-Commerce Application": {
        "type": "http",
        "in": "header",
        "scheme": "bearer"
      }
    }
  }
}